conda-requirements.txt


blas=1.0=mkl
ca-certificates=2020.1.1=0
certifi=2019.11.28=py38_0
cycler=0.10.0=py38_0
freetype=2.9.1=ha9979f8_1
cudatoolkit=10.1.243=h74a9793_0
icc_rt=2019.0.0=h0cc432a_1
icu=58.2=ha66f8fd_1
intel-openmp=2020.0=166
jpeg=9b=hb83a4c4_2

kiwisolver=1.0.1=py38ha925a31_0
libpng=1.6.37=h2a8f88b_0
libtiff=4.1.0=h56a325e_0
matplotlib=3.1.3=py38_0
matplotlib-base=3.1.3=py38h64f37c6_0
mkl=2020.0=166
mkl-service=2.3.0=py38hb782905_0
mkl_fft=1.0.15=py38h14836fe_0
mkl_random=1.1.0=py38hf9181ef_0

ninja=1.9.0=py38h74a9793_0
numpy=1.18.1=py38h93ca92e_0
numpy-base=1.18.1=py38hc3f5095_1
olefile=0.46=py_0
openssl=1.1.1d=he774522_4

pillow=7.0.0=py38hcc1f983_0
pip=20.0.2=py38_1
pyaml=19.4.1=py_0

pyparsing=2.4.6=py_0
pyqt=5.9.2=py38ha925a31_4

python=3.8.1=he1778fa_7_cpython
python-dateutil=2.8.1=py_0
pytorch=1.4.0=py3.8_cuda101_cudnn7_0

pyyaml=5.3=py38he774522_0
qt=5.9.7=vc14h73c81de_0
scipy=1.4.1=py38h9439919_0
setuptools=45.2.0=py38_0

sip=4.19.13=py38ha925a31_0
six=1.14.0=py38_0

sqlite=3.31.1=he774522_0

tk=8.6.8=hfa6e2cd_0

torchvision=0.5.0=py38_cu101
tornado=6.0.3=py38he774522_3

vc=14.1=h0510ff6_4
vs2015_runtime=14.16.27012=hf0eaf9b_1
wheel=0.34.2=py38_0
wincertstore=0.2=py38_0
xz=5.2.4=h2fa13f4_4
yaml=0.1.7=hc54c509_2
zlib=1.2.11=h62dcd97_3
zstd=1.3.7=h508b16e_0
